requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
environs==9.5.0
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
httptools==0.6.0
idna==3.4
kombu==5.3.1
loguru==0.7.0
marshmallow==3.20.1
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
